Refactor test app into modular project and rename folder
