+++ Project_work_Gruppo_4/src/rag_app/__init__.py
@@ -1,0 +1,3 @@
+from .pipeline import RAGPipeline
+
+__all__ = ["RAGPipeline"]
